backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy>=2.0
psycopg2-binary>=2.9
apscheduler>=3.10
requests>=2.31
pyyaml>=6.0
google-auth>=2.28
google-auth-httplib2>=0.2
google-api-python-client>=2.120
gitpython>=3.1
python-dotenv>=1.0
pydantic-settings>=2.2
